# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Dugble Email and SMS API",
    "description": "Send, schedule, inspect, update, and cancel email and SMS messages with Dugble. All endpoints require a team token sent as a Bearer token.",
    "version": "1.0.0",
    "contact": {
      "name": "Dugble Support",
      "email": "support@dugble.com"
    }
  },
  "servers": [
    {
      "url": "https://api.dugble.com",
      "description": "Production"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/emails": {
      "get": {
        "operationId": "listEmails",
        "summary": "List emails",
        "description": "Returns emails belonging to the authenticated team, newest first.",
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/Offset"
          }
        ],
        "responses": {
          "200": {
            "description": "Emails retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailListEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "operationId": "sendEmail",
        "summary": "Send an email",
        "description": "Queues an email for immediate or scheduled delivery. The `to`, `cc`, `bcc`, and `reply_to` fields accept either one address or an array. Addresses may be strings such as `Ada <ada@example.com>` or structured objects.",
        "tags": [
          "Email"
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/IdempotencyKey"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmailSendRequest"
              },
              "example": {
                "from": "Dugble <notifications@example.com>",
                "to": [
                  "Ada <ada@example.com>"
                ],
                "subject": "Welcome to Dugble",
                "html": "<h1>Welcome!</h1><p>Your account is ready.</p>",
                "text": "Welcome! Your account is ready.",
                "tags": [
                  {
                    "name": "category",
                    "value": "welcome"
                  }
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Email accepted for asynchronous delivery.",
            "headers": {
              "Location": {
                "description": "Relative URL of the queued email.",
                "schema": {
                  "type": "string",
                  "example": "/emails/49a3999c-0ce1-4ea6-ab68-afcd6dc2e794"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailAcceptedEnvelope"
                }
              }
            }
          },
          "default": {
            "$ref": "#/components/responses/Error"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "Team token",
        "description": "Pass a Dugble team token in the `Authorization` header as `Bearer <token>`."
      }
    },
    "parameters": {
...
      "Limit": {
        "name": "limit",
        "in": "query",
        "required": false,
        "description": "Maximum number of messages to return. Values outside 1–100 use the default of 50.",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 1,
          "maximum": 100,
          "default": 50
        }
      },
      "Offset": {
        "name": "offset",
        "in": "query",
        "required": false,
        "description": "Number of messages to skip.",
        "schema": {
          "type": "integer",
          "format": "int32",
          "minimum": 0,
          "default": 0
        }
      },
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "A unique key that prevents duplicate mutation requests. Maximum 256 characters.",
        "schema": {
          "type": "string",
          "maxLength": 256
        },
        "example": "order_123_welcome_email"
      }
    },
    "responses": {
      "Error": {
        "description": "The request failed. The HTTP status reflects the error category.",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            },
            "examples": {
              "badRequest": {
                "summary": "Invalid request",
                "value": {
                  "success": false,
                  "error": {
                    "code": "BAD_REQUEST",
                    "message": "Invalid request"
                  }
                }
              },
              "unauthorized": {
                "summary": "Invalid or missing team token",
                "value": {
                  "success": false,
                  "error": {
                    "code": "UNAUTHORIZED",
                    "message": "Authentication is required"
                  }
                }
              }
            }
          }
        }
      }
    },
    "schemas": {
      "EmailAddressObject": {
        "type": "object",
        "required": [
          "email"
        ],
        "properties": {
          "email": {
            "type": "string",
            "format": "email",
            "example": "ada@example.com"
          },
          "name": {
            "type": "string",
            "maxLength": 128,
            "example": "Ada Lovelace"
          }
        },
        "additionalProperties": false
      },
      "EmailAddress": {
        "description": "An email address string, optionally with a friendly name, or a structured address object.",
        "oneOf": [
          {
            "type": "string",
            "example": "Ada Lovelace <ada@example.com>"
          },
          {
            "$ref": "#/components/schemas/EmailAddressObject"
          }
        ]
      },
      "EmailAddressOrList": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/EmailAddress"
          },
          {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmailAddress"
            }
          }
        ]
      },
      "Tag": {
        "type": "object",
        "required": [
          "name",
          "value"
        ],
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "maxLength": 256,
            "example": "category"
          },
          "value": {
            "type": "string",
            "pattern": "^[A-Za-z0-9_-]+$",
            "maxLength": 256,
            "example": "welcome"
          }
        },
        "additionalProperties": false
      },
      "Attachment": {
        "type": "object",
        "required": [
          "filename"
        ],
        "description": "An attachment must provide exactly one of `content` or `path`. Total decoded Base64 content may not exceed 40 MB.",
        "properties": {
          "content": {
            "type": "string",
            "contentEncoding": "base64",
            "description": "Base64-encoded attachment content."
          },
          "filename": {
            "type": "string",
            "minLength": 1,
            "example": "invoice.pdf"
          },
          "path": {
            "type": "string",
            "format": "uri",
            "pattern": "^https?://",
            "example": "https://example.com/invoice.pdf"
          },
          "content_type": {
            "type": "string",
            "example": "application/pdf"
          },
          "content_id": {
            "type": "string",
            "example": "company-logo"
          }
        },
        "oneOf": [
          {
            "required": [
              "content"
            ],
            "properties": {
              "content": {
                "type": "string"
              },
              "path": false
            }
          },
          {
            "required": [
              "path"
            ],
            "properties": {
              "path": {
                "type": "string"
              },
              "content": false
            }
          }
        ],
        "additionalProperties": false
      },
      "EmailSendRequest": {
        "type": "object",
        "required": [
          "to",
          "subject"
        ],
        "properties": {
          "from": {
            "$ref": "#/components/schemas/EmailAddress"
          },
          "to": {
            "$ref": "#/components/schemas/EmailAddressOrList"
          },
          "cc": {
            "$ref": "#/components/schemas/EmailAddressOrList"
          },
          "bcc": {
            "$ref": "#/components/schemas/EmailAddressOrList"
          },
          "reply_to": {
            "$ref": "#/components/schemas/EmailAddressOrList"
          },
          "subject": {
            "type": "string",
            "minLength": 1,
            "maxLength": 255
          },
          "html": {
            "type": "string",
            "maxLength": 1048576,
            "description": "HTML body. At least one of `html` or `text` is required."
          },
          "text": {
            "type": "string",
            "maxLength": 1048576,
            "description": "Plain-text body. At least one of `html` or `text` is required."
          },
          "headers": {
            "type": "object",
            "maxProperties": 100,
            "additionalProperties": {
              "type": "string"
            },
            "example": {
              "X-Entity-Ref-ID": "order-123"
            }
          },
          "attachments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Attachment"
            }
          },
          "tags": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Tag"
            }
          },
          "scheduled_at": {
            "type": "string",
            "description": "Future ISO 8601 date-time or relative value such as `in 5 min`.",
            "examples": [
              "2026-08-05T11:52:01.858Z",
              "in 5 min"
            ]
          },
          "metadata": {
            "type": "object",
            "additionalProperties": true,
            "description": "Custom JSON metadata, up to 16 KB."
          }
        },
        "anyOf": [
          {
            "required": [
              "html"
            ]
          },
          {
            "required": [
              "text"
            ]
          }
        ],
        "additionalProperties": false
      },
...
      "EmailSummary": {
        "type": "object",
        "required": [
          "id",
          "to_email",
          "subject",
          "status",
          "queued_at",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "to_email": {
            "type": "string",
            "format": "email"
          },
          "to_name": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/EmailStatus"
          },
          "provider": {
            "type": "string"
          },
          "queued_at": {
            "type": "string",
            "format": "date-time"
          },
          "submitted_at": {
            "type": "string",
            "format": "date-time"
          },
          "delivered_at": {
            "type": "string",
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "EmailStatus": {
        "type": "string",
        "enum": [
          "queued",
          "processing",
          "submitted",
          "delivered",
          "delayed",
          "bounced",
          "complained",
          "rejected",
          "failed",
          "canceled"
        ]
      },
...
      "ErrorObject": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        },
        "additionalProperties": false
      },
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "$ref": "#/components/schemas/ErrorObject"
          }
        },
        "additionalProperties": false
      },
      "EmailAcceptedEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "$ref": "#/components/schemas/EmailSummary"
          }
        },
        "additionalProperties": false
      },
...
      "EmailListEnvelope": {
        "type": "object",
        "required": [
          "success",
          "data"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EmailSummary"
            }
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}